feat: 搜索选项可搜索挂载目录
